refactor: move actions to /lib
