+++ project/lib/utils/users.ts
@@ -1,5 +1,5 @@
 import { currentUser } from "@clerk/nextjs/server";
-import userQueries from "../db/queries/users";
+import userQueries from "..//queries/users";
 
 export async function getUserId() {
   const user = await currentUser();
